Fix a crash when reconfiguring parameters. (#10)
Essentially the loop to clear "bad" parameters at the end
of parametersUpdate() was crashing when an invalid parameter was
seen.

However, there is really no need to do this.  Instead, what we
do is to ensure that addInputToParameterMap() only really adds
to the parameter map if it was successful.  If it was not
successful, it makes no changes.  This makes the clearing of
the bad part of the map unnecessary, and fixes the crash.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>